fix(spa-editor): split EditorBody and DateBanner into sibling files (PR #549)
